// src/webgl/WebGLExtension.d.ts
import { EnumIO } from "../foundation/misc/EnumIO";
export interface WebGLExtensionEnum extends EnumIO {
}
export declare const WebGLExtension: Readonly<{
    VertexArrayObject: WebGLExtensionEnum;
    TextureFloat: WebGLExtensionEnum;
    TextureHalfFloat: WebGLExtensionEnum;
    TextureFloatLinear: WebGLExtensionEnum;
    TextureHalfFloatLinear: WebGLExtensionEnum;
    InstancedArrays: WebGLExtensionEnum;
}>;